Add
Added the functionality to delete the items from the table and change their priority by using 2 up/down buttons.
For the action buttons I am using emojis (don't judge me, I am on call for 8 days in a row so I'm tired to use font awesome).
Updated priorities and deleted items also update the local storage array of to do items.
Next thing I want to work on is the styling of this to do list single page application.

And I am sorry that I did these changes on the main branch.